Adding Chapter 3 Examples
Reading through Tim Roughgarden's Algorithms Illumninated, and I gave a
try to working through the Chapter 3, Merge and Count split inversions
algolrithm from page 70, 3.2.10.

The code is all in ruby, and the merge_and_count_sort.rb is trying to be
a literal representation of the exercise. merge_and_inverstions is the
full sorting and merging done, but the implementaiton is not taken from
the psuedo code. I translated merge_and_inversions from python. Then
merge_sort.rb was taken from a Honeybadger.io example of merge sort,
which I tweaked a bit.